+++ odorik/__init__.py
@@ -121,6 +121,16 @@ class Odorik(object):
         self._check_response(response)
         return response
 
+    def calls(self, from_date, to_date, line=None):
+        """Returns list of calls."""
+        args = {
+            'from': from_date.isoformat(),
+            'to': to_date.isoformat()
+        }
+        if line is not None:
+            args['line'] = line
+        return self.get_json('calls.json', args)
+
     def callback(self, caller, recipient, line=None):
         """Initiates callback."""
         args = {